Remove unneeded variable initialization
If a variable is initialized with a value unnecessarily, and the code is
changed, the program would start using the previously unused value. That
is not necessarily the behavior expected by the developer making the
change. It would be safer to report use of uninitialized variable and
force the developer to make an explicit fix.